+++ CHANGELOG.md
@@ -33,6 +33,7 @@ Unreleased
 -   Scraped Chibcha (`chb`). (\#430)
 -   Scraped Kashmiri (`kas`). (\#431)
 -   Added Malayalam (`mal`) scrape. (\#434)
+-   Configuration and initial scrape for Dhivehi (`div`, Maldivian). (\#437)
 
 #### Changed